fix some warnings

+++ base/base_bitfield.h
@@ -8,4 +8,4 @@ internal void   bitfield_clear(BitField64* field);
 internal void   bitfield_set(BitField64* field, uint8 index);
 internal bool32 bitfield_is_set(BitField64 field, uint8 index);
 internal bool32 bitfield_is_activated(BitField64 prev, BitField64 next, uint8 index);
-internal int32  bitfield_bitcount(BitField64 v);
+internal uint64 bitfield_bitcount(BitField64 v);
